MessageBox Delete

+++ Apps/Apps/Customer.cs
@@ -115,10 +115,15 @@ namespace Apps
 
             else if (e.ColumnIndex == 1)
             {
-                Database.getInstance().DeleteSupplier(read.Tables[0].Rows[e.RowIndex][0].ToString());
-                read.Tables[0].Rows.RemoveAt(e.RowIndex);
-                tabelCustomer.Update();
-                tabelCustomer.Refresh();
+                DialogResult dialog = MessageBox.Show("Anda yakin ?", "DELETE DATA", MessageBoxButtons.YesNo,MessageBoxIcon.Warning);
+                if (dialog == DialogResult.Yes)
+                {
+                    Database.getInstance().DeleteSupplier(read.Tables[0].Rows[e.RowIndex][0].ToString());
+                    read.Tables[0].Rows.RemoveAt(e.RowIndex);
+                    tabelCustomer.Update();
+                    tabelCustomer.Refresh();
+                }
+                
             }
         }
